--- packages/frontend/app/result/page.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-
-export default function ResultPage() {
-  return (
-    <div>
-      <h1>Result Page</h1>
-      {/* Add your result page content here */}
-    </div>
-  );
-}
-
-
-// sections
-
-// ... existing code ...
