+++ packages/frontend/app/results/page.tsx
@@ -1,10 +1,18 @@
-import type { Metadata } from 'next';
+import type { Metadata, Viewport } from 'next';
 import ResultsPageClient from './ResultsPageClient';
-
-export const metadata: Metadata = {
-  title: '지자체 조회 | 지역자산역량지수(KLACI)',
-};
 
 export default function ResultsPage() {
   return <ResultsPageClient />;
 }
+
+export const metadata: Metadata = {
+  title: 'KLACI',
+  description: 'KLACI 결과 페이지',
+};
+
+export const viewport: Viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 1,
+  userScalable: false,
+};
