+++ packages/frontend/app/summary/page.tsx
@@ -1,10 +1,20 @@
-import type { Metadata } from 'next';
+import type { Metadata, Viewport } from 'next';
 import SummaryPageClient from './SummaryPageClient';
 
-export const metadata: Metadata = {
-  title: '소개 | 지역자산역량지수(KLACI)',
-};
 
 export default function SummaryPage() {
   return <SummaryPageClient />;
 }
+
+
+export const metadata: Metadata = {
+  title: 'KLACI',
+  description: 'KLACI 요약 페이지',
+};
+
+export const viewport: Viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 1,
+  userScalable: false,
+};
