+++ packages/frontend/components/layout/SummaryLayout.tsx
@@ -1,10 +1,19 @@
 import Footer from '@/components/Footer';
 import Header from '@/components/Header/Header';
+import { DARK_MODE_COLORS } from '@/utils/colors';
 import React from 'react';
 
 const SummaryLayout = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div style={{ width: '100%' }}>
+    <div style={{
+      width: '100%',
+      minWidth: '1500px',
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      justifyContent: 'center',
+      backgroundColor: DARK_MODE_COLORS.background,
+    }}>
       <Header />
       <div
         style={{
